Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/WebIDL/ecmascript-binding/default-iterator-object-expected.txt


PASS Default iterator objects for an interface have the same prototype 
PASS Object.prototype.toString returns correct value 
FAIL @@toStringTag has correct value from prototype assert_equals: expected (string) "URLSearchParams Iterator" but got (undefined) undefined
